added hastels tablea in hostels add hostel in settings
