+++ src/ApiData/ContextProvider.js
@@ -1,10 +1,19 @@
-import React, { createContext, useState, useEffect, useContext } from 'react';
+import React, { createContext, useState, useContext, useRef, useEffect } from 'react';
 import { FetchData } from './FetchData';
 import { onValue, ref } from 'firebase/database';
 import { database } from '../firebase';
-
+import isEqual from 'lodash/isEqual';
 
 const DataContext = createContext();
+
+function useDeepCompareEffect(callback, dependencies) {
+  const currentDependenciesRef = useRef();
+  if (!isEqual(currentDependenciesRef.current, dependencies)) {
+    currentDependenciesRef.current = dependencies;
+  }
+  React.useEffect(callback, [currentDependenciesRef.current]);
+}
+
 const DataProvider = ({ children }) => {
   const [data, setData] = useState(null);
   const [activeBoysHostel, setActiveBoysHostel] = useState(null);
@@ -26,10 +35,7 @@ const DataProvider = ({ children }) => {
   }, []);
 
   useEffect(() => {
-    // Reference to the 'Hostel/boys' path in your Firebase database
     const boysRef = ref(database, 'Hostel/boys');
-
-    // Listen for value changes at this reference
     const unsubscribe = onValue(boysRef, (snapshot) => {
       if (snapshot.exists()) {
         const data = snapshot.val();
@@ -40,15 +46,11 @@ const DataProvider = ({ children }) => {
       }
     });
 
-    // Cleanup the listener when the component unmounts
     return () => unsubscribe();
   }, []);
 
   useEffect(() => {
-    // Reference to the 'Hostel/boys' path in your Firebase database
     const girlsRef = ref(database, 'Hostel/girls');
-
-    // Listen for value changes at this reference
     const unsubscribe = onValue(girlsRef, (snapshot) => {
       if (snapshot.exists()) {
         const data = snapshot.val();
@@ -59,28 +61,32 @@ const DataProvider = ({ children }) => {
       }
     });
 
-    // Cleanup the listener when the component unmounts
     return () => unsubscribe();
   }, []);
 
-  useEffect(() => {
+  useDeepCompareEffect(() => {
     if (activeBoysHostelButtons.length > 0) {
       setActiveBoysHostel(activeBoysHostelButtons[0]);
     }
   }, [activeBoysHostelButtons]);
 
-  useEffect(() => {
+  useDeepCompareEffect(() => {
     if (activeGirlsHostelButtons.length > 0) {
       setActiveGirlsHostel(activeGirlsHostelButtons[0]);
     }
   }, [activeGirlsHostelButtons]);
- 
+
+  console.log("active boys hostel", activeBoysHostel);
+  console.log("active buttons", activeBoysHostelButtons);
+
   return (
-    <DataContext.Provider value={{ data, activeBoysHostel, setActiveBoysHostel, activeBoysHostelButtons,  activeGirlsHostel,setActiveBoysHostelButtons,setActiveGirlsHostelButtons, setActiveGirlsHostel, activeGirlsHostelButtons }}>
+    <DataContext.Provider value={{ data, activeBoysHostel, setActiveBoysHostel, activeBoysHostelButtons, activeGirlsHostel, setActiveGirlsHostel, activeGirlsHostelButtons }}>
       {children}
     </DataContext.Provider>
   );
 };
-// Define the useData hook using the DataContext
+
 const useData = () => useContext(DataContext);
 export { DataContext, useData, DataProvider };
+
+
