+++ src/components/DashboardBoys/DashboardBoys.jsx
@@ -1481,6 +1481,9 @@ const DashboardBoys = () => {
         // Logic to open the popup for "Total Beds" card
         setPopupOpen(true);
     }
+    if (item.heading === 'Total Expenses') {
+      setExpensePopupOpen(true);
+    }
   };
